Finalizar reimplementação do login com funcionalidades adicionais

+++ ClientService/src/main/java/loja/clientservice/service/ClientServiceImp.java
@@ -8,6 +8,8 @@ import loja.clientservice.entity.ClientResponse;
 import loja.clientservice.repository.ClientRepository;
 import loja.clientservice.service.mapper.ClientMapper;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.crypto.bcrypt.BCrypt;
 import org.springframework.stereotype.Service;
 
@@ -16,6 +18,7 @@ import java.util.stream.Collectors;
 
 @RequiredArgsConstructor
 @Service
+@Slf4j
 public class ClientServiceImp implements ClientService {
     private final ClientRepository repository;
     private final ClientMapper mapper;
@@ -25,14 +28,20 @@ public class ClientServiceImp implements ClientService {
         return repository.findByUsername(username).map(mapper::toResponse)
                 .orElseThrow(() -> new EntityNotFoundException("Client not found with name " + username));
     }
+
     @Override
     public ClientModel save(ClientRequest request) {
         return repository.save(mapper.toClientModel(request));
     }
 
     public ClientModel login(ClientRequestLogin login) {
-        login.setPassword(BCrypt.hashpw(login.getPassword(), BCrypt.gensalt()));
-        return repository.findByEmailAndPassword(login.getEmail(), login.getPassword());
+        ClientModel client = repository.findByEmail(login.getEmail());
+        if (client != null && checkPassword(login.getPassword(), client.getPassword())){
+            log.info(client.toString());
+            return client;
+        }
+        log.error("Client não esta disponivel");
+        return null;
     }
 
     @Override
@@ -41,8 +50,17 @@ public class ClientServiceImp implements ClientService {
     }
 
     @Override
-    public void delete(ClientRequestLogin login) {
-        ClientModel client = repository.findByEmailAndPassword(login.getEmail(), login.getPassword());
-        repository.delete(client);
+    public HttpStatus delete(ClientRequestLogin login) {
+
+        ClientModel client = repository.findByEmail(login.getEmail());
+        if (client != null && checkPassword(login.getPassword(), client.getPassword())){
+            repository.delete(client);
+            return HttpStatus.OK;
+        }
+        return HttpStatus.NOT_FOUND;
+    }
+
+    public boolean checkPassword(String rawPassword, String securedPassword){
+        return BCrypt.checkpw(rawPassword, securedPassword);
     }
 }
